chore: migrate PreCheckDataRelations/Updatable hook to Validator
migrate
* PreCheckDataRelationsValidationHook
* PreCheckUpdatableFields
to Validator<T>
