+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/config/TrackerValidationConfig.java
@@ -34,12 +34,10 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.hisp.dhis.tracker.validation.TrackerValidationHook;
-import org.hisp.dhis.tracker.validation.hooks.PreCheckDataRelationsValidationHook;
 import org.hisp.dhis.tracker.validation.hooks.PreCheckExistenceValidationHook;
 import org.hisp.dhis.tracker.validation.hooks.PreCheckMandatoryFieldsValidationHook;
 import org.hisp.dhis.tracker.validation.hooks.PreCheckMetaValidationHook;
 import org.hisp.dhis.tracker.validation.hooks.PreCheckUidValidationHook;
-import org.hisp.dhis.tracker.validation.hooks.PreCheckUpdatableFieldsValidationHook;
 import org.hisp.dhis.tracker.validation.hooks.RepeatedEventsValidationHook;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -79,8 +77,6 @@ public class TrackerValidationConfig
             PreCheckExistenceValidationHook.class,
             PreCheckMandatoryFieldsValidationHook.class,
             PreCheckMetaValidationHook.class,
-            PreCheckUpdatableFieldsValidationHook.class,
-            PreCheckDataRelationsValidationHook.class,
 
             /*
              * NB! This hook must be run after all the Event validations,
