+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/validation/hooks/EventPreCheckDataRelationsValidator.java
@@ -27,21 +27,12 @@
  */
 package org.hisp.dhis.tracker.validation.hooks;
 
-import static org.hisp.dhis.tracker.TrackerType.ENROLLMENT;
-import static org.hisp.dhis.tracker.TrackerType.EVENT;
-import static org.hisp.dhis.tracker.TrackerType.TRACKED_ENTITY;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1014;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1022;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1029;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1033;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1041;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1079;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1089;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1115;
 import static org.hisp.dhis.tracker.report.TrackerErrorCode.E1116;
-import static org.hisp.dhis.tracker.report.TrackerErrorCode.E4012;
-import static org.hisp.dhis.tracker.validation.hooks.RelationshipValidationUtils.getUidFromRelationshipItem;
-import static org.hisp.dhis.tracker.validation.hooks.RelationshipValidationUtils.relationshipItemValueType;
 
 import java.util.HashSet;
 import java.util.List;
@@ -59,18 +50,14 @@ import org.hisp.dhis.organisationunit.OrganisationUnit;
 import org.hisp.dhis.program.Program;
 import org.hisp.dhis.program.ProgramInstance;
 import org.hisp.dhis.program.ProgramStage;
-import org.hisp.dhis.trackedentity.TrackedEntityInstance;
-import org.hisp.dhis.tracker.TrackerType;
 import org.hisp.dhis.tracker.bundle.TrackerBundle;
 import org.hisp.dhis.tracker.domain.Enrollment;
 import org.hisp.dhis.tracker.domain.Event;
 import org.hisp.dhis.tracker.domain.MetadataIdentifier;
-import org.hisp.dhis.tracker.domain.Relationship;
-import org.hisp.dhis.tracker.domain.RelationshipItem;
 import org.hisp.dhis.tracker.preheat.TrackerPreheat;
 import org.hisp.dhis.tracker.report.TrackerErrorCode;
-import org.hisp.dhis.tracker.validation.TrackerValidationHook;
 import org.hisp.dhis.tracker.validation.ValidationErrorReporter;
+import org.hisp.dhis.tracker.validation.Validator;
 import org.springframework.stereotype.Component;
 
 /**
@@ -78,36 +65,11 @@ import org.springframework.stereotype.Component;
  */
 @Component
 @RequiredArgsConstructor
-public class PreCheckDataRelationsValidationHook
-    implements TrackerValidationHook
+public class EventPreCheckDataRelationsValidator
+    implements Validator<Event>
 {
     @Override
-    public void validateEnrollment( ValidationErrorReporter reporter, TrackerBundle bundle, Enrollment enrollment )
-    {
-        Program program = bundle.getPreheat().getProgram( enrollment.getProgram() );
-        OrganisationUnit organisationUnit = bundle.getPreheat()
-            .getOrganisationUnit( enrollment.getOrgUnit() );
-
-        reporter.addErrorIf( () -> !program.isRegistration(), enrollment, E1014, program );
-
-        TrackerPreheat preheat = bundle.getPreheat();
-        if ( programDoesNotHaveOrgUnit( program, organisationUnit, preheat.getProgramWithOrgUnitsMap() ) )
-        {
-            reporter.addError( enrollment, E1041, organisationUnit, program );
-        }
-
-        validateTrackedEntityTypeMatchesPrograms( reporter, bundle, program, enrollment );
-    }
-
-    private boolean programDoesNotHaveOrgUnit( Program program, OrganisationUnit orgUnit,
-        Map<String, List<String>> programAndOrgUnitsMap )
-    {
-        return !programAndOrgUnitsMap.containsKey( program.getUid() )
-            || !programAndOrgUnitsMap.get( program.getUid() ).contains( orgUnit.getUid() );
-    }
-
-    @Override
-    public void validateEvent( ValidationErrorReporter reporter, TrackerBundle bundle, Event event )
+    public void validate( ValidationErrorReporter reporter, TrackerBundle bundle, Event event )
     {
         ProgramStage programStage = bundle.getPreheat().getProgramStage( event.getProgramStage() );
         OrganisationUnit organisationUnit = bundle.getPreheat().getOrganisationUnit( event.getOrgUnit() );
@@ -156,6 +118,13 @@ public class PreCheckDataRelationsValidationHook
         {
             reporter.addError( event, E1029, organisationUnit, program );
         }
+    }
+
+    private boolean programDoesNotHaveOrgUnit( Program program, OrganisationUnit orgUnit,
+        Map<String, List<String>> programAndOrgUnitsMap )
+    {
+        return !programAndOrgUnitsMap.containsKey( program.getUid() )
+            || !programAndOrgUnitsMap.get( program.getUid() ).contains( orgUnit.getUid() );
     }
 
     private void validateEventCategoryOptionCombo( ValidationErrorReporter reporter,
@@ -401,76 +370,9 @@ public class PreCheckDataRelationsValidationHook
         return null;
     }
 
-    private void validateTrackedEntityTypeMatchesPrograms( ValidationErrorReporter reporter, TrackerBundle bundle,
-        Program program,
-        Enrollment enrollment )
-    {
-
-        if ( program.getTrackedEntityType() == null )
-        {
-            return;
-        }
-
-        if ( !trackedEntityTypesMatch( bundle, program, enrollment ) )
-        {
-            reporter.addError( enrollment, E1022, enrollment.getTrackedEntity(), program );
-        }
-    }
-
-    private boolean trackedEntityTypesMatch( TrackerBundle bundle, Program program, Enrollment enrollment )
-    {
-        final TrackedEntityInstance trackedEntityInstance = bundle
-            .getPreheat().getTrackedEntity( enrollment.getTrackedEntity() );
-        if ( trackedEntityInstance != null )
-        {
-            return program.getTrackedEntityType().getUid()
-                .equals( trackedEntityInstance.getTrackedEntityType().getUid() );
-        }
-
-        return bundle.findTrackedEntityByUid( enrollment.getTrackedEntity() )
-            .map( te -> te.getTrackedEntityType().isEqualTo( program.getTrackedEntityType() ) )
-            .orElse( false );
-    }
-
-    @Override
-    public void validateRelationship( ValidationErrorReporter reporter, TrackerBundle bundle,
-        Relationship relationship )
-    {
-        validateRelationshipReference( reporter, bundle, relationship, relationship.getFrom() );
-        validateRelationshipReference( reporter, bundle, relationship, relationship.getTo() );
-    }
-
-    private void validateRelationshipReference( ValidationErrorReporter reporter, TrackerBundle bundle,
-        Relationship relationship,
-        RelationshipItem item )
-    {
-        Optional<String> uid = getUidFromRelationshipItem( item );
-        TrackerType trackerType = relationshipItemValueType( item );
-
-        if ( TRACKED_ENTITY.equals( trackerType ) )
-        {
-            if ( uid.isPresent() && !ValidationUtils.trackedEntityInstanceExist( bundle, uid.get() ) )
-            {
-                reporter.addError( relationship, E4012, trackerType.getName(), uid.get() );
-            }
-        }
-        else if ( ENROLLMENT.equals( trackerType ) )
-        {
-            if ( uid.isPresent() && !ValidationUtils.enrollmentExist( bundle, uid.get() ) )
-            {
-                reporter.addError( relationship, E4012, trackerType.getName(), uid.get() );
-            }
-        }
-        else if ( EVENT.equals( trackerType ) && uid.isPresent() && !ValidationUtils.eventExist( bundle, uid.get() ) )
-        {
-            reporter.addError( relationship, E4012, trackerType.getName(), uid.get() );
-        }
-    }
-
     @Override
     public boolean skipOnError()
     {
         return true;
     }
-
 }
